backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
requests>=2.32.3
python-dateutil>=2.9.0.post0
pydantic>=2.7.0
reportlab>=4.2.0
python-dotenv>=1.0.0
